+++ R/bucket_exists.R
@@ -5,6 +5,7 @@
 #'
 #' @return \code{TRUE} if bucket exists and is accessible, else \code{FALSE}.
 #' @references \href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketHEAD.html}{API Documentation}
+#' @seealso \code{\link{bucketlist}}, \code{\link{get_bucket}}, \code{\link{object_exists}}
 #' @export
 bucket_exists <- function(bucket, ...){
     s3HTTP(verb = "HEAD",